refactor : 객체에 메시지를 보내도록 코드 리팩토링
